pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cadet-xray-pipeline"
version = "0.1.0"
description = "Production-ready X-ray cavity detection pipeline built around pycadet."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "OpenAI - generated from user notebook"}
]
dependencies = [
  "numpy>=1.24",
  "matplotlib>=3.8",
  "scipy>=1.11",
  "pandas>=2.1",
  "astropy>=6.0",
  "pyyaml>=6.0",
  "pydantic>=2.7",
  "keras==2.15",
  "tensorflow==2.15",
  "pycadet"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "black>=24.0",
  "ruff>=0.4.0",
  "jupyter>=1.0"
]

[project.scripts]
cadet-xray = "cadet_xray_pipeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
